add an option to electra to load data in non-streaming way
