+++ api-examples/transformer_utils.py
@@ -6,7 +6,7 @@ import baseline.pytorch.embeddings
 import baseline.embeddings
 from baseline.progress import create_progress_bar
 from torch.utils.data.dataset import IterableDataset, TensorDataset
-from baseline.vectorizers import Token1DVectorizer, BPEVectorizer1D
+from baseline.vectorizers import Token1DVectorizer, BPEVectorizer1D, Char2DVectorizer, WordpieceVectorizer1D
 import codecs
 from collections import Counter
 import glob
@@ -251,6 +251,120 @@ class TransformerDiscriminator(nn.Module):
                 real_loss = self.loss(input[target != 0], target[target != 0])
                 return real_loss + fake_loss
         return Loss()
+
+
+class TensorDatasetReaderBase(object):
+    """Provide a base-class to do operations that are independent of token representation
+    """
+    def __init__(self, nctx, vectorizers):
+        self.vectorizers = vectorizers
+        self.nctx = nctx
+        self.num_words = {}
+
+    def build_vocab(self, files):
+        vocabs = {k: Counter({'[CLS]': 1}) for k in self.vectorizers.keys()}
+
+        for file in files:
+            if file is None:
+                continue
+            self.num_words[file] = 0
+            with codecs.open(file, encoding='utf-8', mode='r') as f:
+                sentences = []
+                for line in f:
+                    split_sentence = line.split() + ['<EOS>']
+                    self.num_words[file] += len(split_sentence)
+                    sentences += split_sentence
+                for k, vectorizer in self.vectorizers.items():
+                    vocabs[k].update(vectorizer.count(sentences))
+        return vocabs
+
+    def load_features(self, filename, vocabs):
+
+        features = dict()
+        with codecs.open(filename, encoding='utf-8', mode='r') as f:
+            sentences = []
+            for line in f:
+                sentences += line.strip().split() + ['<EOS>']
+            for k, vectorizer in self.vectorizers.items():
+                vec, valid_lengths = vectorizer.run(sentences, vocabs[k])
+                features[k] = vec[:valid_lengths]
+                shp = list(vectorizer.get_dims())
+                shp[0] = valid_lengths
+                features['{}_dims'.format(k)] = tuple(shp)
+        return features
+
+
+class TensorWordDatasetReader(TensorDatasetReaderBase):
+    """Read each word, and produce a tensor of x and y that are identical
+    """
+    def __init__(self, nctx, use_subword=None, model_file=None, vocab_file=None, special_tokens=None):
+        """Create a reader with a context window that reads words
+
+        :param nctx: The context window length
+        :param use_subword: If this is not none, it should be either 'bpe' or 'wordpiece'
+        """
+        self.use_subword = use_subword
+
+        if self.use_subword == 'bpe':
+            vectorizer = BPEVectorizer1D(model_file=model_file, vocab_file=vocab_file)
+        elif self.use_subword == 'wordpiece':
+            vectorizer = WordpieceVectorizer1D(embed_file=model_file, vocab_file=vocab_file,
+                                               special_tokens=special_tokens)
+        else:
+            vectorizer = Token1DVectorizer(transform_fn=baseline.lowercase)
+        super().__init__(nctx, {'x': vectorizer})
+
+    def build_vocab(self, files):
+        """Read the vocab file to get the tokens
+
+        :param files:
+        :return:
+        """
+        if self.use_subword is not None:
+            super().build_vocab(files)
+            return {'x': self.vectorizers['x'].vocab}
+        return super().build_vocab(files)
+
+    def load(self, filename, vocabs):
+        features = self.load_features(filename, vocabs)
+        x_tensor = torch.tensor(features['x'], dtype=torch.long)
+        num_sequences_word = (x_tensor.size(0) // self.nctx) * self.nctx
+        x_tensor = x_tensor.narrow(0, 0, num_sequences_word).view(-1, self.nctx)
+        return TensorDataset(x_tensor, x_tensor)
+
+
+class TensorCharDatasetReader(TensorDatasetReaderBase):
+    """TensorCharDatasetReader reads in a vocab and then a dataset and returns as a `dict` of `string` to `ndarray`
+    """
+    def __init__(self, nctx, chars_per_word):
+        y_vectorizer = Token1DVectorizer(transform_fn=baseline.lowercase)
+        x_vectorizer = Char2DVectorizer(mxwlen=chars_per_word)
+        super(TensorCharDatasetReader, self).__init__(nctx, {'x': x_vectorizer, 'y': y_vectorizer})
+        self.chars_per_word = chars_per_word
+
+    def load(self, filename, vocabs):
+        features = self.load_features(filename, vocabs)
+        y_tensor = torch.tensor(features['y'], dtype=torch.long)
+        num_sequences_word = (y_tensor.size(0) // self.nctx) * self.nctx
+        y_tensor = y_tensor.narrow(0, 0, num_sequences_word).view(-1, self.nctx)
+
+        x_dataset = torch.tensor(features['x'], dtype=torch.long)
+        x_tensor = torch.tensor(x_dataset, dtype=torch.long)
+        x_tensor = x_tensor.narrow(0, 0, num_sequences_word)
+        x_tensor = x_tensor.view(-1, self.nctx, self.chars_per_word)
+        return TensorDataset(x_tensor, y_tensor)
+
+
+def load_data_caching(token_type, reader, dataset, file_key, vocabs, caching, logger):
+    cached_file = '{}-{}.cache'.format(dataset[file_key], token_type)
+    if caching and os.path.exists(cached_file):
+        logger.info("Reloading %s from cached file [%s]", file_key, cached_file)
+        loaded = torch.load(cached_file)
+    else:
+        loaded = reader.load(dataset[file_key], vocabs)
+        logger.info("Caching %s to [%s]", file_key, cached_file)
+        torch.save(loaded, cached_file)
+    return loaded
 
 
 class MultiFileLoader(IterableDataset):
